Replace `match`-blocks by `if`-blocks for Python <3.10
`match` syntax was introduced with Python 3.10
([PEP 634](https://peps.python.org/pep-0634/)).

We use a [small refactoring tool] to replace `match`-blocks by
`if`-blocks to make the code compatible with Python <3.10.  The code
with `match`-blocks is kept in comments so that these blocks can be
restored when the support for earlier versions of Python will be
dropped.
[small refactoring tool]: (https://github.com/thierry-martinez/match_transformer)