Fix check for density matrix
Reported by Maxime Garnier.
https://github.com/TeamGraphix/graphix/commit/f377001769c2a3ff6ace2a84fdbe7826a6287889#r142237724

The test for Pauli preprocessing in density_matrix.py should be
```
pattern._pauli_preprocessed and input_state != graphix.states.BasicStates.PLUS
```
without `not`.

This bug wasn't caught by `pytest` because the density matrix backend
was never tested with Pauli-preprocessed patterns. I think a natural
place to add such a test is in `test_pattern.py`, by "parametrizing"
`test_pauli_measurment` with `backend`. This was not enough because
`test_pauli_measurment` was defined twice in the same file and the
second definition was hiding the first one. Therefore, this commit:

- rename the first `test_pauli_measurment` into
  `test_pauli_measurement_random_circuit` (fixing the typo by the way);

- "parametrize" `test_pauli_measurement_random_circuit` with
  `backend`, testing `"statevector"` and `"densitymatrix"`;

- a TODO is added for tensor network backend, since the default
  `graph_prep` mode (`auto`) selects the mode `parallel` despite this
  mode is incompatible with non-standardize patterns. I propose to leave
  that for another PR.