Fix #188: Fix Pauli measurement conventions
This commit introduces a new named tuple `graphix.pauli.Measure`
containing a pair `(axis, sign)`, with a static method `try_from`
that returns an instance of `Measure` if the given plane and
angle is a Pauli measure, and `None` otherwise.

`axis` is computed by using the `cos` and `sin` definitions of
`pauli.py`, solving the inconsistency bug reported in #188.

This commit introduces a new `Sign` class to make code manipulating
`Plus` and `Minus` to be more readable than when using `bool`.