Support for `C` commands and domain aliasing in `standardize`.
This commit implements the correct rules for commuting `C` commands.
Moreover, this commit creates new instances of commands instead of
modifying the domains in-place: the behavior is now correct even if
there are aliasing between domains (for instance, if the same instance
of command is added twice or appear in multiple patterns).