Move all Pauli measurements to the front before Pauli presimulation (#209)
This commit adds a method `move_pauli_measurements_to_the_front` that applies
exchange rules to move all Pauli measurements to the front of a
pattern.  This allows all the Pauli measurements to be presimulated,
even if their initial domains depend on other measurements.

On random circuits of 10 qubits (depth 10), number of remaining nodes
after Pauli presimulation is reduced by a factor of 4 in average, and
the time spent on state vector simulation is reduced by a factor
75 (including the time for Pauli presimulation).