Fix #252: pyzx 0.9.0, move to extra, comment graphix-{ibmq,perceval}
This commit moves `pyzx` from `requirements-dev.txt` to
`requirements-extra.txt`, because this is not a development
requirement but an optional dependency.

Tests succeed with pyzx 0.9.0. I am no longer able to reproduce the
errors mentioned in #252 and in
https://github.com/TeamGraphix/graphix/pull/250#issuecomment-2676562457

graphix-ibmq and graphix-perceval are broken (in particular, they are
not installable with Python 3.12). They should be fixed in #261; I
comment them for now.

Dependabot can now track pyzx again.