Enable type-checking for simulator and backends
This commit adds type annotations to the simulator and backend
modules, along with their corresponding tests, enabling full
type-checking with `mypy` while preserving existing functionality.

The separation between `TensorNetworkBackend` and backends that
operate on a full-state representation, such as `StatevecBackend` and
`DensityMatrixBackend`, is now clearer with the introduction of the
abstract classes `FullStateBackend` and `DenseState`, which derive
from `Backend` and `BackendState`, respectively. `StatevecBackend` and
`DensityMatrixBackend` inherit from `FullStateBackend`, while
`Statevec` and `DensityMatrix` inherit from `DenseState`. Note that
the class hierarchy of `BackendState` mirrors that of `Backend`.

The `State` class in `base_backend` has been renamed to `BackendState`
to avoid a name conflict with the `State` class defined in
`graphix.states`, which represents the state of a single qubit.

The classes `Backend[StateT_co]` and
`FullStateBackend[DenseStateT_co]` are now parameterized by covariant
type variables, allowing subclasses to narrow the type of the `state`
field to match their specific state representation. Covariance is
sound in this context because the classes are frozen, and it ensures
that `Backend[BackendState]` is a supertype of all backend classes.

The type `Data` is now declared in a dedicated module, `sim.data`, to
break the import cycle between `base_backend`, `statevec`, and
`density_matrix`. The definition of `Data` requires importing the
`statevec` and `density_matrix` modules, both of which import
`base_backend`; `data` is thus only imported within a type-checking
block in `base_backend`.

An update of `graphix-symbolic` is required to implement `__abs__` in
`Expression`, which is used in the definition of
`get_statevec_norm`. This avoids warnings caused by casting to
`float`, which would discard the imaginary part.