Fix #316: pin package version for `mypy` and `types-networkx` (#318)
This commit pins the package version for `mypy==1.17.0` and
`types-networkx==3.5.0.20250701`. It's a continuation of PR #243 where
we pinned the version of `ruff`.

Pinning these package should make the PR review process smoother.

- In #315, CI broke due to a regression introduced in `mypy==1.17`,
  even though the PR was only about updating `ruff`.

- In #309, CI broke due to stricter typing in newer versions of
  `types-networkx`, which required changes in parts of the code not
  touched by the PR.

By pinning these packages and relying on Dependabot to open dedicated
PRs for updates:

- We avoid unpredicatable CI failures caused by unexpected version
  updates.

- We prevent unrelated changes from being added to PRs just because a
  package update occurred during the review process.

- We get dedicated PRs to manage and test updates.

- We are notified when updates break the CI, even if no new code has
  been pushed since the update was released.

The file `noxfile.yml` is updated to remove Python 3.9 from tested versions for dev, since last versions of `types-networkx` are incompatible with Python 3.9.

The dependencies in `requirements-extra.txt` are now also tested independently of `requirements-dev.txt`.