Fix #328: Add file `py.typed` for marking `graphix` package as typed
This commit adds the file `py.typed` to the `graphix` source tree to
mark the `graphix` package as typed upon installation.

According to the Python documentation [Type information in
libraries](https://typing.python.org/en/latest/spec/distributing.html#packaging-typed-libraries):

> Package maintainers who wish to support type checking of their code
MUST add a marker file named `py.typed` to their package supporting
typing.

This commit adds the CI job `typed-package` in the workflow
`typecheck.yml` to check that `mypy` finds the package when
installed. Note that the first commit of this PR, d1af283, is expected
to fail, in order to verify that the bug described in #328 is
reproducible.

This commit also documents in `CONTRIBUTING.md` that `mypy` and
`pyright` still fail to detect `graphix` when type-checking other
Python projects when `graphix` is installed as an external package,
and documents the workaround, which consists of installing `graphix`
in `compat` or `strict` mode.

```bash
pip install -e . --config-settings editable_mode=strict
```

See https://github.com/pypa/setuptools/issues/3518 for more details.