Fix #173: introduce `StandardizedPattern` class
This commit introduces the `StandardizedPattern` class in the
optimization module. The code from the standardize function has
been moved into this class, allowing access to the decomposed parts
of the standardized pattern through class attributes.

This class enables Pauli preprocessing to avoid relying on
`Pattern.get_measurements_commands` and `Pattern.get_graph`,
since the relevant commands are now extracted during the
standardization process.

Previously, the method checked `Pattern.get_measurements_commands`
to determine whether the pattern was standardized (which was always
true in the context of Pauli preprocessing), and standardized it
if necessary. It then extracted the portion of the pattern consisting
of consecutive M commands. Although this optimization stopped just
after the last M command, checking for standardization still
required iterating over the entire pattern.

This method has now been replaced by `extract_measurement_commands`.
This renaming follows the recommendation in #220 to avoid `get_`
and `set_` method names. The new method produces an iterator over the
M commands, regardless of whether the pattern is standardized.