Revert `BaseM` removed in #205
This reverts the removal of `BaseM`, which was deleted in commit
a07197ae2097403ae0b2d62108b46ef7659a4331.

This commit subsumes #237, which restores `BaseM` but lacks the
associated documentation

Currently, `BaseM` is a type alias for `M`, which is not particularly
useful. However, `BaseM` is used in Veriphix in blind patterns: the
server only knows which node is measured, and the measurement itself
is specified by the client's `MeasureMethod`.

This commit is a step towards upstreaming Graphix's features needed
for Veriphix (and for UBQC in general).