Add OpenQASM 3 exporter for circuits (#343)
This commit introduces an OpenQASM 3 exporter for Graphix
circuits. The functionality was originally proposed in #245 but has
not yet been merged.

The added tests verify that a round-trip through the OpenQASM 3
representation preserves Graphix circuits, using the
graphix-qasm3-parser plugin.  This plugin is therefore
added as a `requirements-dev.txt` dependency.

CI is updated so that `pip install .` can detect the current version
number of Graphix, instead of the default `0.1`: to do so, the whole
history and the tags should be available.  We removed the "-e" option
from CI because it is useless in the CI context.

In the long term the `qasm3_exporter` module will also host the
pattern exporter, but that feature is intentionally omitted from this
PR to keep the change focused.