Enable type-checking for `visualization.py` (#347)
This commit adds type annotations to `visualization.py`, along with
its corresponding tests, enabling full type-checking with mypy while
preserving existing functionality.

`pytest` now uses the `--mpl` plugin to detect any changes (and in particular, regressions) in visualizations.

Pauli nodes are now systematically detected.

There is no longer `save` flag in `Pattern.draw_graph`: the figure is saved to a file as soon as `filename` is not `None` (i.e., contains a `Path`). In this case, no window is opened to show the plot interactively (i.e., the plot is shown interactively in a window only if `filename` is `None`).

**Related issue:**
This PR continues the work started in #302, #308, and #312.
---------
Co-authored-by: matulni <m.uldemolins@gmail.com>