Introduce `BaseN` and `PrepareMethod` (#383)
This commit introduces the classes `BaseN` and `PrepareMethod`, which
allow the user to customize how `N` commands are handled. For
instance, in the VBQC context of Veriphix, `N` commands do not
necessarily prepare the node in the |+> state: for each prepared
qubit, the preparation should be performed by the client, so that the
server does not even see the state.